+++ zetaclient/orchestrator/contextupdater_test.go
@@ -1,16 +1,17 @@
 package orchestrator
 
 import (
-	"context"
 	"testing"
+	"time"
 
 	upgradetypes "cosmossdk.io/x/upgrade/types"
-	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/zeta-chain/node/pkg/chains"
+	"github.com/zeta-chain/node/pkg/ptr"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 	"github.com/zeta-chain/node/zetaclient/testutils/mocks"
+	"github.com/zeta-chain/node/zetaclient/testutils/testlog"
 )
 
 func Test_UpdateAppContext(t *testing.T) {
@@ -24,10 +25,10 @@ func Test_UpdateAppContext(t *testing.T) {
 
 	t.Run("Updates app context", func(t *testing.T) {
 		var (
-			ctx      = context.Background()
-			app      = createAppContext(t, eth, ethParams)
-			zetacore = mocks.NewZetacoreClient(t)
-			logger   = zerolog.New(zerolog.NewTestWriter(t))
+			logger                 = testlog.New(t).Logger
+			chainList, chainParams = parseChainsWithParams(t, eth, ethParams)
+			ctx, app               = newAppContext(t, logger, chainList, chainParams)
+			zetacore               = mocks.NewZetacoreClient(t)
 		)
 
 		// Given zetacore client that has eth and btc chains
@@ -37,6 +38,11 @@ func Test_UpdateAppContext(t *testing.T) {
 			IsInboundEnabled:  true,
 			IsOutboundEnabled: true,
 		}
+		opFlags := observertypes.OperationalFlags{
+			RestartHeight:         123,
+			SignerBlockTimeOffset: ptr.Ptr(time.Second),
+			MinimumVersion:        "",
+		}
 
 		zetacore.On("GetBlockHeight", mock.Anything).Return(int64(123), nil)
 		zetacore.On("GetUpgradePlan", mock.Anything).Return(nil, nil)
@@ -44,6 +50,7 @@ func Test_UpdateAppContext(t *testing.T) {
 		zetacore.On("GetAdditionalChains", mock.Anything).Return(nil, nil)
 		zetacore.On("GetChainParams", mock.Anything).Return(newParams, nil)
 		zetacore.On("GetCrosschainFlags", mock.Anything).Return(ccFlags, nil)
+		zetacore.On("GetOperationalFlags", mock.Anything).Return(opFlags, nil)
 
 		// ACT
 		err := UpdateAppContext(ctx, app, zetacore, logger)
@@ -54,15 +61,18 @@ func Test_UpdateAppContext(t *testing.T) {
 		// New chains should be added
 		_, err = app.GetChain(btc.ChainId)
 		require.NoError(t, err)
+
+		// Check OP flags
+		require.Equal(t, opFlags.RestartHeight, app.GetOperationalFlags().RestartHeight)
 	})
 
 	t.Run("Upgrade plan detected", func(t *testing.T) {
 		// ARRANGE
 		var (
-			ctx      = context.Background()
-			app      = createAppContext(t, eth, ethParams)
-			zetacore = mocks.NewZetacoreClient(t)
-			logger   = zerolog.New(zerolog.NewTestWriter(t))
+			logger                 = testlog.New(t).Logger
+			chainList, chainParams = parseChainsWithParams(t, eth, ethParams)
+			ctx, app               = newAppContext(t, logger, chainList, chainParams)
+			zetacore               = mocks.NewZetacoreClient(t)
 		)
 
 		zetacore.On("GetBlockHeight", mock.Anything).Return(int64(123), nil)